hands_on/blueprints/intento2/requirements.txt
flask
Flask-SQLAlchemy
Flask-Migrate
flask-marshmallow
Flask-Cors
pytest
requests